Miscellaneous fixes for improper method calls and syntax errors.

+++ lib/table_xml_extractor.rb
@@ -1,5 +1,5 @@
 class TableXmlExtractor
-  delegate :document_xml_path, :table_xml_dir, :table_xml_path
+  delegate :document_xml_path, :table_xml_dir, :table_xml_path,
     to: :path_manager
 
   def self.compile(document_numbers, date_str)
@@ -17,23 +17,30 @@ class TableXmlExtractor
   end
 
   def perform
-    file = File.open(document_xml_path)
-    document = Nokogiri::XML(file)
+    if File.exists? document_xml_path("full_text")
+      file = File.open(document_xml_path("full_text"))
+      document = Nokogiri::XML(file)
 
-    tables = document.css('GPOTABLE')
-    Dir.glob("#{table_xml_dir}/*.xml").each{|x| File.unlink(x) }
+      tables = document.css('GPOTABLE')
+      Dir.glob("#{table_xml_dir}/*.xml").each{|x| File.unlink(x) }
 
-    if tables.present?
-      FileUtils.mkdir_p(table_xml_dir)
-      tables.each_with_index do |table_node, i|
-        File.open(table_xml_path(i), 'w'){|f| f.write(table_node.to_xml) }
+      if tables.present?
+        FileUtils.mkdir_p(table_xml_dir)
+        tables.each_with_index do |table_node, i|
+          File.open(table_xml_path(i), 'w'){|f| f.write(table_node.to_xml) }
+        end
       end
+    else
+      Honeybadger.notify(
+        error_class: 'HtmlCompilator::MissingXmlFile',
+        error_message: document_xml_path("full_text")
+      )
     end
   end
 
   private
 
   def path_manager
-    @path_manager ||= XsltPathManager.new(document.document_number, date)
+    @path_manager ||= XsltPathManager.new(document_number, date)
   end
 end
